Further Navigation and Supervisor
If you log in as a TA role, you will not see a button to send notifications. Only viewable if logged in as instructor or supervisor.

On courses page, the "add/delete courses" button is only visible if logged in as supervisor.

Added further functionality to navigating between templates in browser.